+++ sdk/resourcemanager/azure-resourcemanager-storage/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Supported user-assigned service identity for `StorageAccount` in create and update. 
 
+### Other Changes
+
+- Added dependency of `azure-resourcemanager-msi` and `azure-resourcemanager-authorization`.
+
 ## 2.34.0 (2023-12-22)
 
 ### Other Changes